more fault tolerance in decrypt script

+++ decrypt_hw.sh
@@ -1,5 +1,5 @@
 gpg-zip --output homework/my_solutions.tar --decrypt homework/my_solutions.tar.gpg
-if [ $? == 0 ]; then
+if [ $? == 0 ] && [ -e homework/my_solutions.tar ]; then
   rm -R homework/my_solutions.tar.gpg
   tar -xf homework/my_solutions.tar
   rm -R homework/my_solutions.tar
